allow ",optional" to apply to all data types
The optional flag is something we pass through to protoc by copying it
into the protobuf definitions we generate. So it can apply to any type,
not just pointer types.

Also a map is, internally, a pointer type, and you can distinguish
between empty and nil maps. So even without the above choice, map
(and slice) would still need to be handled.

Once "optional" applies to all types, I can simplify how the protobuf
is generated. And finally I added to the unit tests to check a
non-pointer and a map.