fix intermittantly failing unit test
we can't have multiple items in a map and compare the
raw protobuf, because it depends on the [random] map
iteration order. Since map marshaling is already tested
in its own test, I just simplify TestOptional test case
to have only one item in the map, which does not create
any ambiguity when marshaling.