introduce an experimental API for custom marshaling: AppendProtobuf3
which helps some custom marshalers avoid allocating and returning small
slices. It's not that useful because most custom marshalers are for
wiretype bytes, and that encoding requires a varint length prefix,
which requires calculating the size of the encoded value before
encoding, or a memmove() once the true size is discovered.

However for my one case, marshaling a slice of a couple million
structs containing one net.AddrPort4 and one timestamp.NSec (which
both have custom marshalers) along with a few other fields, it sped
marshaling up by 25% and reduced the peak memory usage by not creating
a ton of tiny 6-byte and 8 byte []byte garbage.