Simplified the AppendProtobuf3 API to work like many
other AppendXyz methods in go: take a []byte and return the appended
[]byte or an error.

After implementing several custom marshalers I realized the bytes
wiretype's prefixed length problem made every AppendProtobuf3
method that had to target bytes a mess. So instead there's generic
implementation in this package, and AppendProtobuf3 is not supposed
to encode the byte length.

The other simplification is to split Buffer into Buffer and WriteBuffer,
where WriteBuffer is only the portion of Buffer which a simpler user,
like an AppendProtobuf3 method, might need. It turned out that all
that was needed was []byte.

I think if someday that memmove (currently implemented by an append)
is a performance hotspot then we can add another method by which
this library could query the object what its protobuf length was going
to be, akin to SizeVarint() for varints. But not now.