remove broken and necessary function AppendNSecTimestamp
It's broken if it was used to implement an Appender interface because it
prepended the length of the rest of the protobuf when it should not
have.

It's removed because once that's fixed, it's exactly the same as the
existing EncodeNSecTimestamp() function, so you can use that instead.
And since it's removed, if someone did use it they'll get a code
breakage and hopfully track it down to this commit.

More fun details: why wasn't this caught a long time ago? Because
most of the time the current time is used for unit tests, and the
current time in nanoseconds usually marshals into 11 bytes of
protobuf, so AppendNSecTimestamp would prepend those 11 bytes with
a byte 0x0b. When it came time to unmarshal this, the function
DecodeNSecTimestamp() decodes the 0x0b as a tag|type of some tag
it doesn't understand and calls skip(). skip() return an error
b/c 0x0b is an invalid (in protobuf v3) WireBeginGroup tag. But
DecodeNsecTimestamp() ignored the returned error, and instead
continued on to the next byte, which is the valid 0x08 tag=1,WireVarint,
and from there on decoded the timestamp properly.

The only caused visible trouble when by luck the length wasn't
11, but 10 (and even shorter is possible too, but 10 was bad enough).
That happens when the nanosecond portion of the timestamp has a lot
of leading zeros, so it marshals as a varint into fewer than 4 bytes.
When DecodeNsecTimestamp() gets the 0xa0, that gets interpreted as
a tag=1|Bytes, and we go to skip it too, but then we get off track
because skip() will treat the 0x08 which follows as the length
of the bytes it should skip, and move ahead 8 bytes. At that point
we're decoding data as tags and it's all gone off the rails. The
usual results is a UnexpectedEOF error, but sometimes you'd get so
lucky that the data would result in everything getting skipped without
any overflow, so again DecodeNsecTimestamp() would not return an error,
but on the other hand it wouldn't have found the secs and nanosecs
tags either, so it would return the zero value.

And that zero value is how I began to unravel this bug, just about
2 hours ago.