improve unmarshaling into slices of messages and strings
and [][]byte by preallocating what is probably the correct sized
slice to begin with.

This reducs the number of times append needs to reallocate the slice.
For my (private, large) test case the improvement is

 before:
BenchmarkUnmarshalLargeAPStats-8            3116          11550025 ns/op         3052630 B/op      30315 allocs/op
 after:
BenchmarkUnmarshalLargeAPStats-8            3780           9576556 ns/op         1251903 B/op      10032 allocs/op

That's 1/3 as many allocs (and later gc work), and a 17% performance improvement.

Of course if your data has no slices then this doesn't help.

And if your marshaler didn't put all the elements of a slice (a 'repeated' type) together in the protobuf data
then it doesn't help (but doesn't hurt either). However I have yet to encounter a marshaler that does that.

TODO: in theory this can also be done for []float32 and []float64
easily, since they are encoded as fixed32 and fixed64. To do it for
packed varints would require a custom peek-ahead which counted the
bytes with bit 7 cleared. However since Go does make SIMD easy, doing
this can be as costly.